Generate with text >= 2.0.2 bound
The generated code uses `Data.Text.Internal.pack`, which did not exist
until 2.0.2, so we should also generate this bound.

The generated code should really use `Data.Text.pack` and be compatible
with more versions of `text`, but I am having an impossible time finding
where to make that change.