Support JSON or Yaml specifications
JSON is common, and it's trivial to support as the Yaml parsing is using
`FromJSON` anyway. In addition to guessing by file extension, a
`--spec-format` option was added to be able to specify format for cases
where the file-name can't be used, such when piping `stdin` via a
process substitution. There should probably be first-class support for
`stdin` anyway (`--specification -`), so such a flag will be necessary.